requirements.txt
APScheduler==3.9.1
args==0.1.0
asgiref==3.5.0
Babel==2.10.1
boto3==1.21.46
botocore==1.24.46
certifi==2021.10.8
charset-normalizer==2.0.12
clint==0.5.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.4
django-admin-sortable2==2.0.3
django-anymail==8.5
django-apscheduler==0.6.2
django-checkbox-normalize==0.1.2
django-cleanup==6.0.0
django-colorfield==0.6.3
django-countries==7.3.2
django-currentuser @ git+https://github.com/asucrews/django-currentuser@757dd163ab8906febc6f455a21e427c7fdb90e1b
django-formtools==2.3
django-jet-reboot==1.3.1
django-modeltranslation==0.17.5
django-money==2.1.1
django-phonenumber-field==6.1.0
django-polymorphic==3.1.0
django-referrer-policy==1.0
django-richtextfield==1.6.1
django-solo==2.0.0
django-storages==1.12.3
gopay==1.2.10
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
phonenumbers==8.12.47
Pillow==9.1.0
pip-licenses==3.5.4
psycopg2-binary==2.9.3
PTable==0.9.2
py-moneyed==1.2
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
s3transfer==0.5.2
sentry-sdk==1.5.10
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
whitenoise==6.0.0
wrapt==1.14.0
